requirements.txt
asgiref==3.8.1
contourpy==1.2.1
cycler==0.12.1
Django==5.0.4
django-crispy-forms==2.1
django-phonenumber-field==7.3.0
et-xmlfile==1.1.0
fonttools==4.51.0
intel-openmp==2024.1.0
kiwisolver==1.4.5
mkl==2024.1.0
openpyxl==3.1.2
packaging==24.0
phonenumbers==8.13.35
pillow==10.3.0
psycopg2-binary==2.9.9
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
tbb==2021.12.0
tzdata==2024.1
